Fix test which was missing status workflows
The fixes in 7c7ebb64587 made the spec
`./modules/boards/spec/features/action_boards/custom_field_filters_spec.rb:93`
fail in job run
https://github.com/opf/openproject/actions/runs/17292632687/job/49083361666.

It appears that never really worked because a workflow was missing to be
able to change work package status from open to closed. So moving from
column Open to column Closed would fail. But actually the test was
passing because it was not properly waiting for the column to have
finished updating after moving, so it was ok. Probably it was flaky and
was sometimes failing.

Adding a proper workflow fixed it.