[#67525] API Improvements to `WidgetComponent` and related components (#20350)
* Rework WidgetComponent to render its own WidgetBox

Adds `rows` slots to `WidgetBoxComponent` with the goal of offering a
similar API to `BorderBox` (albeit subset) and obviating the need for
brittle styling overrides.

Also moves responsibility for rendering `turbo-frame` to the Widget Box.

* Remove suffix from nested Header, Body classes

The `Component` suffix is useful in avoiding collisions with other
objects, especially Rails models, but one `Component` in a
full-qualified class name suffices.

* Fix failing test and pass header id from the widget box component

* Create a new component for showing Rows in the body

* Add component_wrapper to be able to act as streamable

---------

Co-authored-by: Behrokh Satarnejad <b.satarnejad@openproject.com>
Co-authored-by: Henriette Darge <h.darge@openproject.com>