Make health check title more generic
The previous title was confusing to read when the
check failed, because the title said "has expected content"
while the description said "has no expected content".

The new title should make it more clear, that the result
is indicated in the description only and that the title
does not indicate a result.