Fix flaky users create spec
failing spec: spec/features/users/create_spec.rb:146
failing run: https://github.com/opf/openproject/actions/runs/18487329720/job/52673204968

I was not able to reproduce the flaky test locally. I still don't know
why it was often failing in CI and never locally.

But I noticed that in the same spec, the one in context "as global user"
had a similar setup and never failed. Probably because it was not using
any browser but the default Capybara driver
`Capybara::RackTest::Driver`.

So the flaky test was changed to not rely on browser-based drivers
anymore. The test setup had to be adapated: as javascript is not
available, the username field is not shown when `ldap_auth_source` is
set, so we create the user first and then update it to set the username.
That's a bit hacky but that works.

This should fix the flakiness of the test.

As a bonus, the test is executed faster now (8 seconds instead of 17).