Extract checking for selection into result method
Adding this method to the result makes it available to any
component involved in the SqlWalker implementation, so it
should be more reusable.

The code that's calculating the actual SELECT SQL statements is
not making use of the new helper, so some duplication still exists,
but hopefully we'll not deduplicate more than the current state.