Fix CSP issue in backlogs module
https://community.openproject.org/wp/64158

Inline javascript inside html header does not work when loaded after a
turbo visit because the nonce is removed by turbo when it loads the
page, and then the browser CSP blocks the inline javascript from the header
because the nonce is missing. This happens on Firefox.

Fix is to use the `content_for :additional_js_dom_ready` hook to load
the javascript instead.